Move windhelm window lights and reduce brightness
Some windhelm windows are used in the military barracks in Ark. They're
really low to the floor and quite bright in close proximity, which was
causing overexposure in this location in Enderal.

Reduce brightness of these window lights and move them higher, so
they're level with the windows.

This will probably make some floors darker in Skyrim, but I'm not sure
which ones. Shouldn't matter if you're using Window Shadows Ultimate
though.